python-sanic-psyconpg/requirements.txt
sanic
uvicorn[standard]
psycopg
psycopg_pool
python-dotenv
black
